Add event click to hour
- Motivation
Would be good if we have and event to retrieve date and time without
events just in case we need or want to create an event using that empty
space.

- Solution
Add a function `hourTapped` that will receive our date and time
selected, with this date and time selected we can add a new event.
In `DayView` was changed the current `View` to a `TouchableOpacity` with
a `View` inside to mark space.